gfx/layers/apz/test/mochitest/test_click.html

<!DOCTYPE HTML>
<html>
<head>
  <meta charset="utf-8">
  <title>Sanity clicking test</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
  <script type="application/javascript">

SimpleTest.waitForExplicitFinish();

// this page just serially loads each one of the following test helper pages in
// a new window and waits for it to call testDone()
var tests = [
  // Sanity test to synthesize a mouse click
  {'file': 'helper_click.html'},
  // Sanity test for click but with some mouse movement between the down and up
  {'file': 'helper_drag_click.html'}
];

var testIndex = -1;
var w = null;

function testDone() {
  var test = tests[testIndex];
  if (w) {
    if (typeof test.dp_suppression != 'undefined') {
      // We modified the suppression when starting the test, so now undo that.
      SpecialPowers.getDOMWindowUtils(window).respectDisplayPortSuppression(!test.dp_suppression);
    }
    if (!!test.prefs) {
      // We pushed some prefs for this test, pop them, and re-invoke
      // testDone() after that's been processed
      SpecialPowers.popPrefEnv(function() {
        w.close();
        w = null;
        testDone();
      });
      return;
    }

    w.close();
  }

  testIndex++;
  if (testIndex >= tests.length) {
    SimpleTest.finish();
    return;
  }

  test = tests[testIndex];
  if (typeof test.dp_suppression != 'undefined') {
    // Normally during a test, the displayport will get suppressed during page
    // load, and unsuppressed at a non-deterministic time during the test. The
    // unsuppression can trigger a repaint which interferes with the test, so
    // to avoid that we can force the displayport to be unsuppressed for the
    // entire test which is more deterministic.
    SpecialPowers.getDOMWindowUtils(window).respectDisplayPortSuppression(test.dp_suppression);
  }
  if (!!test.prefs) {
    // Got some prefs for this subtest, push them
    SpecialPowers.pushPrefEnv({"set": test.prefs}, function() {
      w = window.open(test.file, "_blank");
    });
  } else {
    w = window.open(test.file, "_blank");
  }
}

window.onload = function() {
  if (!SpecialPowers.getDOMWindowUtils(window).asyncPanZoomEnabled) {
    ok(true, "APZ is not enabled, this test is not relevant, sorry!\n");
    SimpleTest.finish();
    return;
  }
  testDone();
};

  </script>
</head>
<body>
</body>
</html>
